Cambio a BigDecimal de los valores para cálculo. Se añaden como parámetros la precisión, número de decimales a mostrar el resultada y el RoundingMode, la forma de redondeo de esa precición

+++ src/main/java/org/math/calculator/infraestructure/rest/response/OperationResult.java
@@ -4,9 +4,11 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 @ToString
 @Getter
 @Builder
 public class OperationResult {
-    private Double result;
+    private BigDecimal result;
 }
